+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format will be:
 
 ### Refactored
 - Minimized code for fetching the favicon in the `/favicon.ico` endpoint.
+- Changed implementation of loading environment variables through `pydantic-settings` instead of `python-dotenv`.
 
 ### Performance
 - Changed the implementation of rate limiting to Redis via Upstash (Vercel).